+++ app/Models/User.php
@@ -2,7 +2,6 @@
 
 namespace App\Models;
 
-// use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
@@ -11,14 +10,8 @@ use Illuminate\Support\Str;
 
 class User extends Authenticatable
 {
-    /** @use HasFactory<\Database\Factories\UserFactory> */
     use HasFactory, Notifiable;
 
-    /**
-     * The attributes that are mass assignable.
-     *
-     * @var list<string>
-     */
     protected $fillable = [
         'name',
         'email',
@@ -38,11 +31,6 @@ class User extends Authenticatable
         'last_login',
     ];
 
-    /**
-     * The attributes that should be hidden for serialization.
-     *
-     * @var list<string>
-     */
     protected $hidden = [
         'password',
         'remember_token',
@@ -50,12 +38,6 @@ class User extends Authenticatable
         'discord_refresh_token',
         'github_token',
     ];
-
-    /**
-     * The attributes that should be cast.
-     *
-     * @var array<string, string>
-     */
     protected $casts = [
         'email_verified_at' => 'datetime',
         'password' => 'hashed',
@@ -75,17 +57,5 @@ class User extends Authenticatable
     public function isMod(): bool
     {
         return $this->role === self::ROLE_MOD;
-    }
-
-    public function makeAdmin(string $email): bool
-    {
-        $user = User::where('email', $email)->first();
-        
-        if (!$user) {
-            return false;
-        }
-        
-        $user->role = self::ROLE_ADMIN;
-        return $user->save();
     }
 }